+++ lib/plugins/aws/package/compile/events/httpApi.js
@@ -4,7 +4,7 @@ const _ = require('lodash');
 const d = require('d');
 const memoizee = require('memoizee');
 const memoizeeMethods = require('memoizee/methods');
-const { logWarning } = require('../../../../../../classes/Error');
+const { logWarning } = require('../../../../../classes/Error');
 
 const allowedMethods = new Set(['GET', 'POST', 'PUT', 'PATCH', 'OPTIONS', 'HEAD', 'DELETE']);
 const methodPattern = new RegExp(`^(?:\\*|${Array.from(allowedMethods).join('|')})$`, 'i');
